chore: refactor leaderboard into its own module
